Expose refcount mechanism publicly and use it for freeing.
This allows code to make use of reference counting to keep track of
message and bundle object ownership.  Since lo_message_free(),
lo_bundle_free() and lo_bundle_free_recursive() decrement the counter
and check it before freeing, previous code will work as expected,
however if lo_message_incref() or lo_bundle_incref() are used, actual
free will be deferred until no more references exist.

User code should call always the corresponding lo_message_incref() or
lo_bundle_incref() function when these objects are stored.

+++ src/lo_types_internal.h
@@ -82,7 +82,7 @@ typedef struct _lo_message {
     lo_arg **argv;
     /* timestamp from bundle (LO_TT_IMMEDIATE for unbundled messages) */
     lo_timetag ts;
-		size_t refcount;
+    int refcount;
 } *lo_message;
 
 typedef int (*lo_method_handler) (const char *path, const char *types,
@@ -174,8 +174,8 @@ struct _lo_bundle {
     size_t size;
     size_t len;
     lo_timetag ts;
-		lo_element *elmnts;
-		size_t refcount;
+    lo_element *elmnts;
+    int refcount;
 };
 
 typedef struct _lo_strlist {
